inline: CM-885 convert static method to subs
See http://blogs.perl.org/users/rurban/2011/06/how-perl-calls-subs-and-methods.html
It is better to do it here with the inline peep check, and not in ck_subr.

But we cannot use the $obj->method type info, because $obj may be a subtype
of the declared type, which may have a different method.