iter: relink range and nested loopvar WIP
global $_ loopvar or refs are nested: rv2gv -> gv _
with `for $_ ()`, or `for \$ref ()`.

Also handle missing a loop var when a lexical.

Note that a normal scalar symbol itervar $sv or $_ must be always
rv2gv - gv, as enteriter/cx_pushloop_for can only handle gv or pad
itervars, a sv is always a magic ref.
This should be changed to allow fast non-lexical GVSV itervars.
Only implicit $_ is pretty fast this way, as it does not need a rv2gv deref.

- This passes all t/op/for.t tests, but crashes with make_patchnum
 at pp_iter_ary: corrupt cx->blk_loop.state_u.ary.ary
 in a nested loop.

nested loops:
SEGV:
 gmpb -Dtsk -e'@a=(0..1);for my $a (@a) { print $_ for (@a) }'
panic:
 gmpb -Dtxk -e'my @names=qw(1 2);for my $n (@names) { for my $x (keys %ENV){print $x}}'