ffi: skip miniperl, set ENTERFFI type
handle GvCV, SvRV and direct CV in ENTERFFI.
no overloaded object and no LV.

Set the right type already in the ck_subr to have less
op misses in entersub, also for xsub calls.
Note that a compile-time ffi call will not fall back to
a user-sub if redefined. Only done for XS to Perl or Perl to XS/FFI.

Simplify a few builtins.

Change attribute args: "native(" is followed by the arg.