ffi: magic ffienc for the :encoded attribute WIP
for string arguments and return type, PAD sv attachments.

Maybe use it also for all string values (as in Perl6), added to Encode when
being used outside of Latin-1/UTF-8.  But then as index into an encoded
array of names.  Now as shared name, attached as new F ffienc magic.