ffi: fix sub EXISTING_SYM () :native :symbol(OTHERSYM)
use flat lists as attrs argument to attributes->import, not
nested. this was ignore in dup_attrslist().
add a testcase to test for the 2nd :symbol(OTHER) to resolve, not
skip.
Change the :symbol is already resolved warning to redefine, and only
if it's superfluous, resolving to the same symbol.