ffi: c++ -fpermissive compat
use memcpy() which does a better copy of the wanted size,
and is optimized to the same code as a simple cast.
esp. we don't use a union, when we can use a single and more precise line.

we cannot use void* with C++, use char* instead.

improve the tests to use int, not long.