ffi: fix GVSV threaded
On threads a GVSV for the native($n) argument is invalid,
we need to use SVREF(GVOP) which is a PADOP.
This failed in op destruction on the invalid padix.

There is a problem with the padix from the upper scope.
Use the previous padix when copying it to the attributes->import
call.